Monitoring DAQ (#186)
* Added small Monitor DAQ tool

* Added error messages to MonitorDAQ tool

* Update README of MonitorDAQ tool

* fixed slack connections after tests

* Added boolean config variable for sending warning messages to slack

* Remove delete call for context, ServiceDiscovery

* Convert UTC time to Fermilab time in MonitorDAQ

* added beam/cosmic separation in MonitorMRDTime --> better representation of beam neutrinos

* added memory & disk space monitoring to MonitorDAQ

* changed Path to CompStats files

* Set default colors for text in infoboxes (MonitorDAQ)

Set the default color to black for text in infoboxes in the MonitorDAQ tool (particularly for the memory and disk space monitoring)

* Added clear calls for infobox canvases

Two TCanvas::Clear() calls were added in the PrintInfoBox() function since the infobox canvases were sometimes showing a plot in the background.

* RWM buffer plot is now averaged over multiple waveforms

* added sanity checks for memory consumption and cpu values

* added disk monitoring

* added averaged BRF signal

* Minor update for current time in InfoBox

* Updated inactive MRD channel list

* temporary update of the loopback configuration

* Add file size check to MonitorReceive

* Added Above100 variable

* Added Above100 to MonitorDAQ

* Changed file size limit to 1.8GB

* Changed file size limit to 1.3GB

* Changed the file size limit to 1.0GB

* Updated file size limit to 2GB

* Added try/catch phrases to MonitorReceive

MonitorReceive would crash in some rare cases where there is no MRDData object in the raw data file. The try/catch phrases seem to prevent this crash from happening

* Removed verbosity variable from MonitorReceive tool

MonitorReceive was expecting to have a verbosity integer variable since the last commit that introduced the try/catch phrases. However, no such verbosity variable was declared in its header file, so all instances of verbosity are replaced with `0` in the MonitorReceive tool in this commit.

* Prevent slack warnings for disk space bug

* Updated dead channel list for Monitoring (MRD)

* Updated dead channel list for Monitoring (PMT)

* Change threshold of entries from 14000 to 2000 for MonitorReceive tool

* Changed file size limit from 100MB to 50MB