mprove consistency of the CNNImage tool by adding a "Static" mode that positions PMTs based on a static mapping / lookup table.
"Static" is now the preferred mode to use the tool with. In the previous (and currently still maintained) "PMT-Wise" mode, the position of the PMTs in the resulting CNNImage is re-calculated at each execution of the event loop. This method is susceptible to minor changes in the PMT positions, see for example the correction factors in the new CNNImage::OrderPMTPositions method. The step toward a fixed mapping will therefore increase stability and grant continuity.

As a further step, some code cleanup has begun and compartmentalization of the code into smaller methods is also under way. The goal of adding further methods is to decrease clutter and increase readability.

The README file and config files of respectively the UserTools/CNNImage and configfiles/CNNImage has been changed to reflect some of the changes.

The mapping is generated / explained with a few plots in the added python notebook, which uses the PMTPositions_Scan.txt. The lines within those plots indicate the binning borders.