Update ChannelSPEGains2023.csv
this channel does not have a gains measurement as its off/inactive. However, for some reason in very rare instances a hit is registered (not sure why this is the case). Set the gains to 0 in case a hit is registered, as to assign it 0 ADC value and not contribute any charge.