Fix a bug about NumWavesInCompleteSet in EBPMT
Previously, if NumWavesInCompleteSet was set to 140 and there are only ~133 PMT waveforms exist in data, the AlmostCompleteWaveforms[PMTCounterTimeNs] won't be created. But the usage of AlmostCompleteWaveforms[PMTCounterTimeNs] in Log will create an entry there. This won't affect the later code because there are checks to require AlmostCompleteWaveforms[PMTCounterTimeNs] > 5.

Now the Log was removed and the tolerance for PMT number is changed to 10 at:
ChannelKey.size() >= (NumWavesInCompleteSet - 10)

Also change the pedestal subtraction for LAPPD_EB tool chain to be 0.