Fixing bug in PhaseIITreeMaker.cpp (#348)
* Fixing bug in PhaseIITreeMaker.cpp

NDigitsPMTs and NDigitsLAPPDs are initialized to -9999. Then, they are incremented in order to count the number of PMT/LAPPD digits in the event. But this way, there is a -9999 offset to their values. So I initialize them to 0 before incrementing. Also I removed the totalPMTs and totalLAPPDs variables because they are basically not used.

* Fixing typo

* Update PhaseIITreeMaker.cpp

Initialize NDigitsPMTs and NDigitsLAPPDs variables in ResetVariables() function.