Add conversion from ACCID to Incom ID for runs before 5887, Also add new BRF channel in PMTDataDecoder
The new Incom ID is saved in the PsecData during data taken, and this
happends after run 5887.  The information filling in ANNIEEventTreeMaker use ID from the PsecData, so
for newer runs after 5887, the Incom ID is automatically filled.

For older runs before 5887, add the conversion for LAPPD_ID.

The BRF signal is also changed to new slot, use run 5887 as the changing
point. However there were some test runs at the end of beam year 2025,
they are ignored.